+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* Version 2 configurations for ISO 19115-1 and ISO 19115-2 to support backwards incompatible schema changes
+* Version 3 configurations for ISO 19115-1 and ISO 19115-2 to support backwards incompatible schema changes
   [#182](https://gitlab.data.bas.ac.uk/uk-pdc/metadata-infrastructure/metadata-library/-/issues/182)
 * Documentation on how to add a new configuration version for an existing standard
   [#196](https://gitlab.data.bas.ac.uk/uk-pdc/metadata-infrastructure/metadata-library/-/issues/196)
+* Permissions configuration property within resource constraints element
+  [#191](https://gitlab.data.bas.ac.uk/uk-pdc/metadata-infrastructure/metadata-library/-/issues/191)
 
 ### Changed